Move forward function out of base layer
